services/api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = "AudioMind API Gateway"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "slowapi>=0.1.9",
    "PyJWT>=2.9",
    "bcrypt>=4.0",
    "pydantic-settings>=2.6",
    # Redis
    "redis[hiredis]>=5.0",
    # OpenTelemetry
    "opentelemetry-api>=1.28",
    "opentelemetry-sdk>=1.28",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-exporter-otlp-proto-http>=1.28",
]

# Virtual project: this service is not distributed as a library.
# uv installs its dependencies into the workspace virtualenv but does not
# build or install the service itself as a Python package.
[tool.uv]
package = false
